requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
mysql-connector-python==8.2.0
numpy==1.26.2
pandas==2.1.4
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2023.4
urllib3==2.1.0
yarl==1.9.4
PyYAML~=6.0.1